services/main-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
google-generativeai
python-dotenv
redis

# Development and testing
ruff
pytest
pytest-cov
httpx # For testing FastAPI endpoints
requests
